+++ src/Planners/tests/UnicycleTrajectoryGeneratorTest.cpp
@@ -32,10 +32,13 @@ std::shared_ptr<IParametersHandler> params()
     handler->setParameter("leftContactFrameName", "l_sole");
     handler->setParameter("rightContactFrameName", "r_sole");
 
+    // Override some default parameters
+    handler->setParameter("nominalWidth", 0.2);
+
     return handler;
 }
 
-TEST_CASE("UnicycleTrajectoryGeneratorTest")
+TEST_CASE("Simple run", "[UnicycleTrajectoryGenerator]")
 {
 
     auto jointsList
@@ -57,6 +60,138 @@ TEST_CASE("UnicycleTrajectoryGeneratorTest")
     BipedalLocomotion::Planners::UnicycleTrajectoryGenerator unicycleTrajectoryGenerator;
 
     REQUIRE(unicycleTrajectoryGenerator.initialize(handler));
+
+    REQUIRE(unicycleTrajectoryGenerator.setRobotContactFrames(ml.model()));
+
+    UnicycleTrajectoryGeneratorInput input
+        = UnicycleTrajectoryGeneratorInput::generateDummyUnicycleTrajectoryGeneratorInput();
+
+    UnicycleTrajectoryGeneratorOutput output;
+
+    REQUIRE(unicycleTrajectoryGenerator.setInput(input));
+    REQUIRE(unicycleTrajectoryGenerator.advance());
+    REQUIRE(unicycleTrajectoryGenerator.isOutputValid());
+
+    output = unicycleTrajectoryGenerator.getOutput();
+
+    REQUIRE(output.contactPhaseList.size() == 1);
+}
+
+TEST_CASE("Initialization with Feet Realitve Pose", "[UnicycleTrajectoryGenerator]")
+{
+
+    auto jointsList
+        = std::vector<std::string>{"l_hip_pitch",      "l_hip_roll",       "l_hip_yaw",
+                                   "l_knee",           "l_ankle_pitch",    "l_ankle_roll",
+                                   "r_hip_pitch",      "r_hip_roll",       "r_hip_yaw",
+                                   "r_knee",           "r_ankle_pitch",    "r_ankle_roll",
+                                   "torso_pitch",      "torso_roll",       "torso_yaw",
+                                   "neck_pitch",       "neck_roll",        "neck_yaw",
+                                   "l_shoulder_pitch", "l_shoulder_roll",  "l_shoulder_yaw",
+                                   "l_elbow",          "r_shoulder_pitch", "r_shoulder_roll",
+                                   "r_shoulder_yaw",   "r_elbow"};
+
+    iDynTree::ModelLoader ml;
+    REQUIRE(ml.loadReducedModelFromFile(getRobotModelPath(), jointsList));
+
+    const auto handler = params();
+
+    BipedalLocomotion::Planners::UnicycleTrajectoryGenerator unicycleTrajectoryGenerator;
+
+    auto leftToRightTransform = manif::SE3d::Identity();
+    double nominalWidth{};
+    handler->getParameter("nominalWidth", nominalWidth);
+    leftToRightTransform.translation(Eigen::Vector3d(0.0, -nominalWidth, 0.0));
+
+    REQUIRE(unicycleTrajectoryGenerator.initialize(handler, leftToRightTransform));
+
+    REQUIRE(unicycleTrajectoryGenerator.setRobotContactFrames(ml.model()));
+
+    UnicycleTrajectoryGeneratorInput input
+        = UnicycleTrajectoryGeneratorInput::generateDummyUnicycleTrajectoryGeneratorInput();
+
+    UnicycleTrajectoryGeneratorOutput output;
+
+    REQUIRE(unicycleTrajectoryGenerator.setInput(input));
+    REQUIRE(unicycleTrajectoryGenerator.advance());
+    REQUIRE(unicycleTrajectoryGenerator.isOutputValid());
+
+    output = unicycleTrajectoryGenerator.getOutput();
+
+    REQUIRE(output.contactPhaseList.size() == 1);
+}
+
+TEST_CASE("Initialization with Base Position", "[UnicycleTrajectoryGenerator]")
+{
+
+    auto jointsList
+        = std::vector<std::string>{"l_hip_pitch",      "l_hip_roll",       "l_hip_yaw",
+                                   "l_knee",           "l_ankle_pitch",    "l_ankle_roll",
+                                   "r_hip_pitch",      "r_hip_roll",       "r_hip_yaw",
+                                   "r_knee",           "r_ankle_pitch",    "r_ankle_roll",
+                                   "torso_pitch",      "torso_roll",       "torso_yaw",
+                                   "neck_pitch",       "neck_roll",        "neck_yaw",
+                                   "l_shoulder_pitch", "l_shoulder_roll",  "l_shoulder_yaw",
+                                   "l_elbow",          "r_shoulder_pitch", "r_shoulder_roll",
+                                   "r_shoulder_yaw",   "r_elbow"};
+
+    iDynTree::ModelLoader ml;
+    REQUIRE(ml.loadReducedModelFromFile(getRobotModelPath(), jointsList));
+
+    const auto handler = params();
+
+    BipedalLocomotion::Planners::UnicycleTrajectoryGenerator unicycleTrajectoryGenerator;
+
+    Eigen::Vector3d basePosition = Eigen::Vector3d(0.0, 0.0, 0.7);
+
+    REQUIRE(unicycleTrajectoryGenerator.initialize(handler, basePosition));
+
+    REQUIRE(unicycleTrajectoryGenerator.setRobotContactFrames(ml.model()));
+
+    UnicycleTrajectoryGeneratorInput input
+        = UnicycleTrajectoryGeneratorInput::generateDummyUnicycleTrajectoryGeneratorInput();
+
+    UnicycleTrajectoryGeneratorOutput output;
+
+    REQUIRE(unicycleTrajectoryGenerator.setInput(input));
+    REQUIRE(unicycleTrajectoryGenerator.advance());
+    REQUIRE(unicycleTrajectoryGenerator.isOutputValid());
+
+    output = unicycleTrajectoryGenerator.getOutput();
+
+    REQUIRE(output.contactPhaseList.size() == 1);
+}
+
+TEST_CASE("Initialization with Base Position and Feet Relative Pose",
+          "[UnicycleTrajectoryGenerator]")
+{
+
+    auto jointsList
+        = std::vector<std::string>{"l_hip_pitch",      "l_hip_roll",       "l_hip_yaw",
+                                   "l_knee",           "l_ankle_pitch",    "l_ankle_roll",
+                                   "r_hip_pitch",      "r_hip_roll",       "r_hip_yaw",
+                                   "r_knee",           "r_ankle_pitch",    "r_ankle_roll",
+                                   "torso_pitch",      "torso_roll",       "torso_yaw",
+                                   "neck_pitch",       "neck_roll",        "neck_yaw",
+                                   "l_shoulder_pitch", "l_shoulder_roll",  "l_shoulder_yaw",
+                                   "l_elbow",          "r_shoulder_pitch", "r_shoulder_roll",
+                                   "r_shoulder_yaw",   "r_elbow"};
+
+    iDynTree::ModelLoader ml;
+    REQUIRE(ml.loadReducedModelFromFile(getRobotModelPath(), jointsList));
+
+    const auto handler = params();
+
+    BipedalLocomotion::Planners::UnicycleTrajectoryGenerator unicycleTrajectoryGenerator;
+
+    auto leftToRightTransform = manif::SE3d::Identity();
+    double nominalWidth{};
+    handler->getParameter("nominalWidth", nominalWidth);
+    leftToRightTransform.translation(Eigen::Vector3d(0.0, -nominalWidth, 0.0));
+
+    Eigen::Vector3d basePosition = Eigen::Vector3d(0.0, 0.0, 0.7);
+
+    REQUIRE(unicycleTrajectoryGenerator.initialize(handler, basePosition, leftToRightTransform));
 
     REQUIRE(unicycleTrajectoryGenerator.setRobotContactFrames(ml.model()));